Add contact section to
Need to refactor contact controller and views to make it work with
HTML and JS requests.

Reuse contact form everywhere.

Add translations to Spanish and English.
