+++ app/controllers/budget_controller.rb
@@ -1,4 +1,5 @@
 class BudgetController < ApplicationController
   def new
+    @contact = Contact.new
   end
 end
