--- app/views/contact/greetings.html.haml
@@ -1,4 +1,0 @@
-.page.greetings
-  .cont-introduction.medium-12.small-12.large-8.text-center.row
-    %h1 <strong>Gracias </strong> por contactar con nosotros.
-    %p.text-center Obtendrá una respuesta en menos de 24 horas.
