--- app/views/contact/new.html.haml
@@ -1,10 +1,0 @@
-.page.contact
-  .row
-    .small-12.large-6.small-centered.columns.cont-introduction
-      %h1 Gracias por su interés en los servicios de RockAndRor
-      %p.text-center Escríbanos, le contestaremos en menos de 24 horas.
-      = form_for @contact, url:contact_index_path do |f|
-        = f.text_field :name, placeholder: 'Nombre o Empresa'
-        = f.text_field :email, placeholder: 'Email'
-        = f.text_area :message, placeholder: 'Mensaje'
-        = f.submit "Contacta ahora", class: "large button expand"
