+++ app/views/contacts/show.html.haml
@@ -1,0 +1,4 @@
+.page.greetings
+  .cont-introduction.medium-12.small-12.large-8.text-center.row
+    %h1= t(".title_html")
+    %p.text-center= t(".description")
