+++ app/views/shared/_navbar_animate.html.haml
@@ -84,6 +84,8 @@
     %ul.left.small-4.medium-4.columns.show-for-medium-up
       %li.anim{ class: active?("welcome", "home")}
         %a{ href: root_path }= t("navbar.home")
+      %li.anim.consul{class: active?("budget", "new")}
+        %a{ href: consul_budget_path }= t("navbar.consul")
       %li.anim.has-dropdown.portfolio{ class: active?("welcome", ["boatjump", "webdings", "eatbooking", "oasiscatamaran"]) }
         %a{ href: "#" }= t("navbar.works")
         %ul.dropdown
@@ -95,7 +97,5 @@
             = link_to "Eatbooking", "/eatbooking"
           %li{class: active?("welcome", "oasiscatamaran") }
             = link_to "Oasis Catamaran", "/oasiscatamaran"
-      %li.anim.consul{class: active?("budget", "new")}
-        %a{ href: consul_budget_path }= t("navbar.consul")
       %li.anim.consul{class: active?("contact", ["new", "create"]) }
         %a{ href: new_contact_path }= t("navbar.contact")
