+++ app/views/shared/contact/_form.html.haml
@@ -1,0 +1,5 @@
+= form_for contact, url: contacts_path, remote: remote do |f|
+  = f.text_field :name, placeholder: t("helpers.placeholder.contact.name"), label: false
+  = f.text_field :email, placeholder: t("helpers.placeholder.contact.email"), label: false
+  = f.text_area :message, placeholder: t("helpers.placeholder.contact.message"), rows: 7, label: false
+  = f.submit t(".submit"), class: "large button expand success"
