+++ config/locales/activemodel.en.yml
@@ -1,0 +1,13 @@
+en:
+  activemodel:
+    attributes:
+      contact:
+        name: Name
+        email: Email
+        message: Message
+  helpers:
+    placeholder:
+      contact:
+        name: Enter your name or your institution's name
+        email: Valid email address
+        message: Enter your message
