+++ config/locales/budget.en.yml
@@ -41,3 +41,5 @@ en:
           name: Sebastià Roig Pieras
         senen:
           name: Senén Rodero Rodríguez
+      contact:
+        title: Need more information or an appointment?
