+++ config/locales/budget.es.yml
@@ -41,3 +41,5 @@ es:
           name: Sebastià Roig Pieras
         senen:
           name: Senén Rodero Rodríguez
+      contact:
+        title: ¿Necesita más información o una cita?
