+++ config/locales/en.yml
@@ -13,3 +13,14 @@ en:
     contact: Contact us
     call_us: Call us
     consul: Consul Services
+  contacts:
+    new:
+      title: Thank you for your interest in RockAndRor's services
+      description: Write to us, we will answer you in less than 24 hours.
+    show:
+      title_html: <strong>Thanks </strong> for contacting us.
+      description: You will get an answer in less than 24 hours.
+  shared:
+    contact:
+      form:
+        submit: "Send message"
