+++ config/routes.rb
@@ -5,8 +5,8 @@ Rails.application.routes.draw do
   get '/webdings', to: 'welcome#webdings', as: :webdings
   get '/eatbooking', to: 'welcome#eatbooking', as: :eatbooking
   get '/oasiscatamaran', to: 'welcome#oasiscatamaran', as: :oasiscatamaran
-  resources :contact, only: [:new,:create]
-  get '/greetings', to: 'contact#greetings', as: :greetings
+  resources :contacts, only: [:new, :create]
+  get '/greetings', to: 'contacts#show', as: :greetings
   get '/consul/budget', to: 'budget#new', as: :consul_budget
 
   if Rails.env.development?
